integration_tests/requirements.txt
requests==2.25.1
pytest==5.1.2             # via -r test.in, pytest-datadir, pytest-flask-sqlalchemy, pytest-mock
pytest-datadir==1.3.0     # via -r test.in
pytest-flask-sqlalchemy==1.0.2  # via -r test.in
pytest-mock==1.10.4       # via pytest-flask-sqlalchemy
python-decouple==3.3
python-decouple==3.3
psycopg2-binary==2.8.6
